+++ tinymce/static/django_tinymce/init_tinymce.js
@@ -53,18 +53,24 @@
     }
   }
 
+  function initializeTinyMCE(element, formsetName) {
+    Array.from(element.querySelectorAll('.tinymce')).forEach(area => initTinyMCE(area));
+  }
+
   ready(function() {
     // initialize the TinyMCE editors on load
-    document.querySelectorAll('.tinymce').forEach(function(el) {
-      initTinyMCE(el);
-    });
+    initializeTinyMCE(document);
 
     // initialize the TinyMCE editor after adding an inline in the django admin context.
     if (typeof(django) !== 'undefined' && typeof(django.jQuery) !== 'undefined') {
-      django.jQuery(document).on('formset:added', function(event, $row, formsetName) {
-        $row.find('textarea.tinymce').each(function() {
-          initTinyMCE(this);
-        });
+      django.jQuery(document).on('formset:added', (event, $row, formsetName) => {
+        if (event.detail && event.detail.formsetName) {
+          // Django >= 4.1
+          initializeTinyMCE(event.target);
+        } else {
+          // Django < 4.1, use $row
+          initializeTinyMCE($row.get(0));
+        }
       });
     }
   });